LayoutTests/storage/indexeddb/factory-deletedatabase-interactions-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test the deleteDatabase call and its interaction with open/setVersion

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "factory-deletedatabase-interactions.html"

TEST: deleteDatabase blocked on open handles
self.testname = 'test1'; self.ver = 1
'h.open'
'h.open.onsuccess'
'deleteDatabase()'
'h.onversionchange'
    in versionchange, old = 1 new = ""
    h closing, but not immediately
'deleteDatabase().onblocked'
'h.close'
'deleteDatabase().onsuccess'
PASS blockedEventFired is true

TEST: deleteDatabase not blocked when handles close immediately
self.testname = 'test2'; self.ver = 1
'h.open'
'h.open.onsuccess'
'deleteDatabase()'
'h.onversionchange'
    in versionchange, old = 1 new = ""
    h closing immediately
'h.close'
'deleteDatabase().onblocked'
'deleteDatabase().onsuccess'
NOTE: Will FAIL with extra bogus deleteDatabase().onblocked step; https://bugs.webkit.org/show_bug.cgi?id=71130
FAIL blockedEventFired should be false. Was true.

TEST: deleteDatabase is delayed if a VERSION_CHANGE transaction is running
self.testname = 'test3'; self.ver = 1
'h.open'
'h.open.onsuccess'
'h.setVersion'
'deleteDatabase()'
'h.setVersion.onsuccess'
'h.setVersion.transaction-complete'
'h.onversionchange'
    in versionchange, old = "1" new = ""
    h closing, but not immediately
'deleteDatabase().onblocked'
'h.close'
'deleteDatabase().onsuccess'
PASS versionChangeComplete is true

TEST: Correct order when there are pending setVersion, delete and open calls.
self.testname = 'test4'; self.ver = 1
'h.open'
'h.open.onsuccess'
'h2.open'
'h2.open.onsuccess'
'h.setVersion'
'deleteDatabase()'
'h2.onversionchange'
    in versionchange, old = 1 new = "1"
'h.setVersion.onblocked'
'h3.open'
'h2.close'
'h.setVersion.onsuccess'
'h.setVersion.transaction-complete'
'h.onversionchange'
    in versionchange, old = "1" new = ""
    h closing, but not immediately
'deleteDatabase().onblocked'
'h.close'
'deleteDatabase().onsuccess'
'h3.open.onsuccess'
PASS versionChangeComplete is true
PASS blockedEventFired is true
PASS deleteDatabaseComplete is true
PASS successfullyParsed is true

TEST COMPLETE
